qa_agent_plugin/backend/requirements.txt
# QA Agent - Agentic Architecture v2.0
# Python Dependencies

# Core Framework
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.10.2

# LLM Integration
google-genai==1.10.0

# Model Context Protocol (MCP)
mcp==1.3.1

# Utilities
python-dotenv==1.0.1
rich==13.9.4

# Additional Dependencies
httpx==0.28.1
anyio==4.7.0

